Simplify project structure (part 2).
