+++ package.json
@@ -15,8 +15,8 @@
     "prettier": "prettier --write --list-different '**/*.{ts,js?(on),md,yml}'",
     "prettier-check": "yarn run prettier --write=false",
     "eslint": "eslint '**/*.ts?(x)'",
-    "build": "tsc -b .",
-    "test": "nyc --reporter=lcov mocha 'template/**/*.test.ts'",
+    "build": "yarn --cwd template build",
+    "test": "yarn --cwd template test",
     "coverage": "codecov",
     "deduplicate": "yarn-deduplicate -s fewer",
     "generate": "yarn -s && yarn run ts-node -P tsconfig.json ./scripts/generate.ts",
